Remove the `cached_entries` from the `ReentrantCollectionView`. (#4628)
## Motivation

Having a cache can be a good thing. But of course it can go OOM.
Ideally, an LRU cache should keep track of its memory usage.
Next best would be to have an LRU cache that keeps track of the number
of entries
or something related to the memory usage.
The `cached_entries` of the `ReentrantCollectionView` had none of those
things.

## Proposal

The `cached_entries` is removed altogether from the
`ReentrantCollectionView`.
This is not a problem. The caching should be done in only one place, and
that's in the LRU cache.

## Test Plan

The CI. No change of API from this work.

## Release Plan

It is fully possible to backport this change to the TestNet / DevNet.

## Links

None.